requirements.txt
flask>=2.3.0
pandas>=2.0.0
openpyxl>=3.1.0
pyxlsb>=1.0.10
numpy>=1.24.0
google-generativeai>=0.7.0
pypdf>=4.0.0
python-dotenv
gunicorn
flask-cors
python-pptx>=0.6.21
supabase
xlrd>=2.0.1
